+++ src/Piece/PieceForm/PieceForm.test.js
@@ -321,6 +321,8 @@ describe('PieceForm', () => {
         isCreateAnother: false,
         isCreateItem: false,
         receivingStatus: PIECE_STATUS.expected,
+        itemId: 'itemUUID',
+        request: {},
       },
     });