+++ src/common/hooks/useHoldingsAndLocations/index.js
@@ -1,0 +1,1 @@
+export { useHoldingsAndLocations } from './useHoldingsAndLocations';